address cgets's comments
