+++ ambry-cloud/src/main/java/com.github.ambry.cloud/CloudBackupManager.java
@@ -11,32 +11,35 @@
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  */
-package com.github.ambry.replication;
+package com.github.ambry.cloud;
 
 import com.codahale.metrics.MetricRegistry;
 import com.github.ambry.clustermap.ClusterMap;
-import com.github.ambry.clustermap.VirtualCloudReplicatorClusterMap;
+import com.github.ambry.clustermap.VirtualReplicatorClusterMap;
 import com.github.ambry.config.CloudConfig;
 import com.github.ambry.config.ClusterMapConfig;
 import com.github.ambry.config.ReplicationConfig;
 import com.github.ambry.config.StoreConfig;
 import com.github.ambry.network.ConnectionPool;
 import com.github.ambry.notification.NotificationSystem;
+import com.github.ambry.replication.ReplicationEngine;
+import com.github.ambry.replication.ReplicationException;
 import com.github.ambry.store.StoreKeyConverterFactory;
 import com.github.ambry.store.StoreKeyFactory;
 import java.util.concurrent.ScheduledExecutorService;
 
 
 /**
- * {@link BackupManager} is used to backup partitions to Cloud. Partitions assignment is handled by Helix.
+ * {@link CloudBackupManager} is used to backup partitions to Cloud. Partitions assignment is handled by Helix.
  */
-public class BackupManager extends ReplicationEngine {
+public class CloudBackupManager extends ReplicationEngine {
 
-  public BackupManager(CloudConfig cloudConfig, ReplicationConfig replicationConfig, ClusterMapConfig clusterMapConfig,
-      StoreConfig storeConfig, StoreKeyFactory storeKeyFactory, ClusterMap clusterMap,
-      VirtualCloudReplicatorClusterMap virtualCloudReplicatorClusterMap, ScheduledExecutorService scheduler,
-      ConnectionPool connectionPool, MetricRegistry metricRegistry, NotificationSystem requestNotification,
-      StoreKeyConverterFactory storeKeyConverterFactory, String transformerClassName) throws ReplicationException {
+  public CloudBackupManager(CloudConfig cloudConfig, ReplicationConfig replicationConfig,
+      ClusterMapConfig clusterMapConfig, StoreConfig storeConfig, StoreKeyFactory storeKeyFactory,
+      ClusterMap clusterMap, VirtualReplicatorClusterMap virtualCloudReplicatorClusterMap,
+      ScheduledExecutorService scheduler, ConnectionPool connectionPool, MetricRegistry metricRegistry,
+      NotificationSystem requestNotification, StoreKeyConverterFactory storeKeyConverterFactory,
+      String transformerClassName) throws ReplicationException {
 
     super(replicationConfig, clusterMapConfig, storeKeyFactory, clusterMap, scheduler,
         virtualCloudReplicatorClusterMap.getDataNodeId(clusterMapConfig.clusterMapHostName,
