+++ ambry-replication/src/main/java/com.github.ambry.replication/DiskTokenPersistor.java
@@ -31,7 +31,7 @@ import org.slf4j.LoggerFactory;
  */
 class DiskTokenPersistor implements ReplicaTokenPersistor {
 
-  private Logger logger = LoggerFactory.getLogger(getClass());
+  private final Logger logger = LoggerFactory.getLogger(getClass());
   private final short version = 0;
   private final String replicaTokenFileName;
   private final Map<String, List<PartitionInfo>> partitionGroupedByMountPath;
